FEniCS backend: Move projection code
