+++ tlm_adjoint/fenics/__init__.py
@@ -10,8 +10,8 @@ from .backend_code_generator_interface import linear_solver  # noqa: E402,F401
 from .backend_interface import *   # noqa: E402,F401
 from .caches import *              # noqa: E402,F401
 from .equations import *           # noqa: E402,F401
-from .fenics_equations import *    # noqa: E402,F401
 from .functions import *           # noqa: E402,F401
 from .interpolation import *       # noqa: E402,F401
+from .projection import *          # noqa: E402,F401
 
 from .backend_patches import *     # noqa: E402,F401
